+++ spec/helpers/admin/claims_helper_spec.rb
@@ -99,7 +99,7 @@ describe Admin::ClaimsHelper do
     end
 
     context "when the claim is approaching its deadline" do
-      let(:claim) { create(:claim, :submitted, submitted_at: 5.weeks.ago) }
+      let(:claim) { create(:claim, :submitted, submitted_at: (Claim::CHECK_DEADLINE - 1.week).ago) }
 
       it "always includes the deadline date" do
         expect(helper.admin_submission_details(claim)[2].last).to have_content(l(claim.check_deadline_date))
@@ -154,14 +154,14 @@ describe Admin::ClaimsHelper do
     after { travel_back }
 
     context "when a claim is approaching it's deadline" do
-      let(:claim) { build(:claim, :submitted, submitted_at: 5.weeks.ago) }
+      let(:claim) { build(:claim, :submitted, submitted_at: (Claim::CHECK_DEADLINE - 1.week).ago) }
 
       it { is_expected.to have_content("7 days") }
       it { is_expected.to have_selector(".tag--warning") }
     end
 
     context "when a claim has passed it's deadline" do
-      let(:claim) { build(:claim, :submitted, submitted_at: 10.weeks.ago) }
+      let(:claim) { build(:claim, :submitted, submitted_at: (Claim::CHECK_DEADLINE + 4.weeks).ago) }
 
       it { is_expected.to have_content("-28 days") }
       it { is_expected.to have_selector(".tag--alert") }